make recording golint compliant
