--- recording/constants.go
@@ -1,6 +1,0 @@
-package recording
-
-const (
-	getURL  = "https://api.twilio.com/2010-04-01/Accounts/%s/Recordings/%s.json"
-	listURL = "https://api.twilio.com/2010-04-01/Accounts/%s/Recordings.json"
-)
