+++ recording/recording.go
@@ -4,35 +4,61 @@ import (
 	"errors"
 	"fmt"
 	"github.com/natebrennand/twiliogo/common"
-	"net/http"
 	"net/url"
 	"regexp"
 )
 
+// holds url values used in queries
+var recordings = struct {
+	Get, List string
+}{
+	Get:  "/2010-04-01/Accounts/%s/Recordings/%s.json",
+	List: "/2010-04-01/Accounts/%s/Recordings.json",
+}
+
 var validateRecordingSid = regexp.MustCompile(`^RE[0-9a-z]{32}$`).MatchString
 
+// Account wraps the common Account struct to embed the AccountSid & Token.
 type Account struct {
-	AccountSid string
-	Token      string
-	Client     http.Client
+	common.Account
 }
 
-func (act Account) GetSid() string {
-	return act.AccountSid
-}
-func (act Account) GetToken() string {
-	return act.Token
-}
-func (act Account) GetClient() http.Client {
-	return act.Client
+// Recording represents a voice recording from a phone call.
+type Recording struct {
+	Sid         string          `json:"sid"`
+	DateCreated common.JSONTime `json:"date_created"`
+	DateUpdated common.JSONTime `json:"date_updated"`
+	AccountSid  string          `json:"account_sid"`
+	CallSid     string          `json:"call_sid"`
+	Duration    int64           `json:"duration,string"`
+	APIVersion  string          `json:"api_version"`
+	URI         string          `json:"uri"`
 }
 
+// Get returns data about recording as json.
+// Can get .mp3 or .wav of recording from the uri provided in Recording
+func (act Account) Get(recSid string) (Recording, error) {
+	var r Recording
+	if !validateRecordingSid(recSid) {
+		return r, errors.New("Invalid sid")
+	}
+	err := common.SendGetRequest(fmt.Sprintf(recordings.Get, act.AccountSid, recSid), act, &r)
+	return r, err
+}
+
+// List holds a list of Reording references.
+type List struct {
+	common.ListResponseCore
+	Recordings *[]Recording `json:"recordings"`
+}
+
+// ListFilter can be used to filter the results of a list recordings query.
 type ListFilter struct {
 	CallSid     string
 	DateCreated *common.JSONTime
 }
 
-func (f ListFilter) GetQueryString() string {
+func (f ListFilter) getQueryString() string {
 	v := url.Values{}
 	if f.CallSid != "" {
 		v.Set("CallSid", f.CallSid)
@@ -47,28 +73,18 @@ func (f ListFilter) GetQueryString() string {
 	return encoded
 }
 
-// Returns data about recording as json
-// Can get .mp3 or .wav of recording from the uri provided in Recording
-func (act Account) Get(recSid string) (Recording, error) {
-	var r Recording
-	if !validateRecordingSid(recSid) {
-		return r, errors.New("Invalid sid")
-	}
-	err := common.SendGetRequest(fmt.Sprintf(getURL, act.AccountSid, recSid), act, &r)
-	return r, err
-}
-
-func (act Account) List(f ListFilter) (RecordingList, error) {
-	var rl RecordingList
-	err := common.SendGetRequest(fmt.Sprintf(listURL, act.AccountSid)+f.GetQueryString(), act, &rl)
+// List returns a list of recordings that fulfill the provided filter.
+func (act Account) List(f ListFilter) (List, error) {
+	var rl List
+	err := common.SendGetRequest(fmt.Sprintf(recordings.List, act.AccountSid)+f.getQueryString(), act, &rl)
 	return rl, err
 }
 
-// Returns data about recording as json
+// Delete removes a recording from Twilio servers.
 // Can get .mp3 or .wav of recording from the uri provided in Recording
 func (act Account) Delete(recSid string) error {
 	if !validateRecordingSid(recSid) {
 		return errors.New("Invalid sid")
 	}
-	return common.SendDeleteRequest(fmt.Sprintf(getURL, act.AccountSid, recSid), act)
+	return common.SendDeleteRequest(fmt.Sprintf(recordings.Get, act.AccountSid, recSid), act)
 }
